# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Coach Sutra - Coaching Institute Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      summary: List all batches
      parameters:
        - name: courseId
          in: query
          schema:
            type: string
        - name: academicYear
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
    post:
      operationId: createBatch
      tags: [batches]
      summary: Create a batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}:
    get:
      operationId: getBatch
      tags: [batches]
      summary: Get a batch with students
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
    patch:
      operationId: updateBatch
      tags: [batches]
      summary: Update a batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
    delete:
      operationId: deleteBatch
      tags: [batches]
      summary: Delete a batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /batches/{id}/students:
    post:
      operationId: addStudentToBatch
      tags: [batches]
      summary: Add student to batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchStudentInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"

  # Timetable
...
  /attendance/student:
    get:
      operationId: listStudentAttendance
      tags: [attendance]
      summary: List student attendance records
      parameters:
        - name: batchId
          in: query
          schema:
            type: string
        - name: studentId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: markStudentAttendance
      tags: [attendance]
      summary: Mark student attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/staff:
    get:
      operationId: listStaffAttendance
      tags: [attendance]
      summary: List staff attendance records
      parameters:
        - name: staffId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffAttendanceRecord"
    post:
      operationId: markStaffAttendance
      tags: [attendance]
      summary: Mark staff attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffAttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAttendanceRecord"
  /attendance/student/summary:
    get:
      operationId: getStudentAttendanceSummary
      tags: [attendance]
      summary: Get attendance summary for a student
      parameters:
        - name: studentId
          in: query
          required: true
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"

  # Homework
...
  /admissions:
    get:
      operationId: listAdmissions
      tags: [admissions]
      summary: List admission enquiries
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdmissionEnquiry"
    post:
      operationId: createAdmission
      tags: [admissions]
      summary: Submit admission enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionEnquiry"
  /admissions/{id}:
    patch:
      operationId: updateAdmission
      tags: [admissions]
      summary: Update admission enquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionEnquiry"

  # PTM
...
components:
  schemas:
...
    SignupInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, teacher, staff]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string

    # Dashboard
...
    Batch:
      type: object
      required: [id, name, courseId, capacity, schedule, academicYear, startDate, status]
      properties:
        id:
          type: string
        name:
          type: string
        courseId:
          type: string
        courseName:
          type: ["string", "null"]
        capacity:
          type: integer
        enrolled:
          type: integer
        schedule:
          type: string
        academicYear:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, upcoming]
        studentIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
    BatchInput:
      type: object
      required: [name, courseId, capacity, schedule, academicYear, startDate]
      properties:
        name:
          type: string
        courseId:
          type: string
        capacity:
          type: integer
        schedule:
          type: string
        academicYear:
          type: string
        startDate:
          type: string
        endDate:
          type: string
    BatchUpdate:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: integer
        schedule:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, completed, upcoming]
    BatchStudentInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: string

    # Timetable
...
    AttendanceRecord:
      type: object
      required: [id, studentId, batchId, date, status]
      properties:
        id:
          type: string
        studentId:
          type: string
        studentName:
          type: ["string", "null"]
        batchId:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    AttendanceInput:
      type: object
      required: [studentId, batchId, date, status]
      properties:
        studentId:
          type: string
        batchId:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late]
        remarks:
          type: string
    StaffAttendanceRecord:
      type: object
      required: [id, staffId, date, status]
      properties:
        id:
          type: string
        staffId:
          type: string
        staffName:
          type: ["string", "null"]
        date:
          type: string
        status:
          type: string
          enum: [present, absent, leave]
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    StaffAttendanceInput:
      type: object
      required: [staffId, date, status]
      properties:
        staffId:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [present, absent, leave]
        checkIn:
          type: string
        checkOut:
          type: string
        remarks:
          type: string
    AttendanceSummary:
      type: object
      required: [studentId, totalClasses, present, absent, late, percentage]
      properties:
        studentId:
          type: string
        totalClasses:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
        percentage:
          type: number

    # Homework
...
    AdmissionEnquiry:
      type: object
      required: [id, studentName, parentName, phone, email, courseInterest, status]
      properties:
        id:
          type: string
        studentName:
          type: string
        parentName:
          type: string
        phone:
          type: string
        email:
          type: string
        courseInterest:
          type: string
        courseName:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, visited, enrolled, dropped]
        followUpDate:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    AdmissionInput:
      type: object
      required: [studentName, parentName, phone, email, courseInterest]
      properties:
        studentName:
          type: string
        parentName:
          type: string
        phone:
          type: string
        email:
          type: string
        courseInterest:
          type: string
        source:
          type: string
          enum: [website, referral, social-media, walk-in, other]
        followUpDate:
          type: string
        remarks:
          type: string
    AdmissionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, visited, enrolled, dropped]
        followUpDate:
          type: string
        remarks:
          type: string

    # PTM
...